Refactor: Moved Settings and Cache to own folders.
