+++ src/gsudo/Commands/KillCacheCommand.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
+using gsudo.CredentialsCache;
 
 namespace gsudo.Commands
 {
