+++ src/gsudo/CredentialsCache/CacheMode.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel;
+
+namespace gsudo.CredentialsCache
+{
+    enum CacheMode
+    {
+        Disabled,
+        Explicit,
+        [Description("Enabling the credentials cache is a security risk.")]
+        Auto,
+        //        [Description("Enabling the credentials cache is a security risk.")]
+        //        Unsafe,
+    }
+}
