+++ src/gsudo/Settings/RegistrySetting.cs
